--- vike-svelte/renderer/types.ts
@@ -1,16 +1,0 @@
-import { ComponentType as Component } from "svelte"
-
-export type { PageProps }
-export type { Component }
-
-type PageProps = Record<string, unknown>
-
-declare global {
-  namespace Vike {
-    interface PageContext {
-      Page: Component
-      pageProps: PageProps
-      title?: string
-    }
-  }
-}
