+++ vike-svelte/renderer/utils/isCallable.js
@@ -1,0 +1,8 @@
+/**
+ * @template { (...args: unknown[]) => unknown } T
+ * @param { T | unknown } thing 
+ * @returns {thing is T}
+ */
+export function isCallable(thing) {
+  return thing instanceof Function || typeof thing === 'function'
+}
