--- vike-svelte/renderer/utils/isCallable.ts
@@ -1,3 +1,0 @@
-export function isCallable<T extends (...args: unknown[]) => unknown>(thing: T | unknown): thing is T {
-  return thing instanceof Function || typeof thing === 'function'
-}
